Day 2, part 1

+++ src/y2018/day1/Part1.java
@@ -2,18 +2,18 @@ package y2018.day1;
 
 import utils.Problem;
 
-public class Part1 extends Problem<Long> {
+public class Part1 extends Problem<Integer> {
     private Parser parser;
 
-    public Part1(String fileName, Long exp) {
+    public Part1(String fileName, int exp) {
         super(exp);
 
         parser = new Parser(fileName);
     }
 
-    public Long run() {
+    public Integer run() {
         var ints = parser.parse();
-        var total = 0L;
+        var total = 0;
 
         for (var i : ints) {
             total += i;
